Move commands into own namespace
