+++ bin/up.php
@@ -7,7 +7,7 @@ $app = new \Symfony\Component\Console\Application('up');
 foreach (\Smrtr\MysqlVersionControl\DbConfig::getEnvironments() as $env) {
 
     $app->add(
-        new \Smrtr\MysqlVersionControl\UpCommand($env)
+        new \Smrtr\MysqlVersionControl\Command\UpCommand($env)
     );
 }
