+++ src/Smrtr/MysqlVersionControl/Command/UpCommand.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Smrtr\MysqlVersionControl;
+namespace Smrtr\MysqlVersionControl\Command;
 
+use Smrtr\MysqlVersionControl\DbConfig;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
